modules/lidar/requirements.txt
azure-iothub-device-client~=1.4.3
pigpio
smbus
RPi.GPIO